requirements.txt
streamlit
pandas
numpy
matplotlib
altair
rake-nltk
scikit-learn
networkx
torch
dgl
